refactor: extract TLongMultiIterator for reuse and conciseness
